rename pdf preview component
